+++ index.html
@@ -231,5 +231,9 @@
   </template>
 
 </body>
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
+<script>
+function q(){
+  alert("Hello JavaScript!");
+}</script>
 </html>
